+++ src/elements/widgets/designerView/extensions/TransformOriginExtension.ts
@@ -18,9 +18,9 @@ export class TransformOriginExtension extends AbstractExtension {
     const rect = this.extendedItem.element.getBoundingClientRect();
     const computed = getComputedStyle(this.extendedItem.element);
     const to = computed.transformOrigin.split(' ');
-    this._circle = this._drawCircleOverlay(rect.x - this.designerView.containerBoundingRect.x + Number.parseInt(to[0].replace('px', '')), rect.y - this.designerView.containerBoundingRect.y + Number.parseInt(to[1].replace('px', '')), 5, 'svg-transform-origin');
+    this._circle = this._drawCircleOverlay(rect.x - this.designerView.containerBoundingRect.x + Number.parseFloat(to[0].replace('px', '')), rect.y - this.designerView.containerBoundingRect.y + Number.parseFloat(to[1].replace('px', '')), 5, 'svg-transform-origin');
     this._circle.setAttribute('style', 'cursor: pointer');
-    this._circle2 = this._drawCircleOverlay(rect.x - this.designerView.containerBoundingRect.x + Number.parseInt(to[0].replace('px', '')), rect.y - this.designerView.containerBoundingRect.y + Number.parseInt(to[1].replace('px', '')), 1, 'svg-transform-origin');
+    this._circle2 = this._drawCircleOverlay(rect.x - this.designerView.containerBoundingRect.x + Number.parseFloat(to[0].replace('px', '')), rect.y - this.designerView.containerBoundingRect.y + Number.parseFloat(to[1].replace('px', '')), 1, 'svg-transform-origin');
     this._circle2.setAttribute('style', 'pointer-events: none');
     this._circle.addEventListener(EventNames.PointerDown, event => this.pointerEvent(event));
     this._circle.addEventListener(EventNames.PointerMove, event => this.pointerEvent(event));
@@ -43,10 +43,10 @@ export class TransformOriginExtension extends AbstractExtension {
         if (this._startPos && event.buttons > 0) {
           const dx = event.x - this._startPos.x;
           const dy = event.y - this._startPos.y;
-          this._circle.setAttribute('cx', <any>(rect.x - this.designerView.containerBoundingRect.x + Number.parseInt(to[0].replace('px', '')) + dx));
-          this._circle.setAttribute('cy', <any>(rect.y - this.designerView.containerBoundingRect.y + Number.parseInt(to[1].replace('px', '')) + dy));
-          this._circle2.setAttribute('cx', <any>(rect.x - this.designerView.containerBoundingRect.x + Number.parseInt(to[0].replace('px', '')) + dx));
-          this._circle2.setAttribute('cy', <any>(rect.y - this.designerView.containerBoundingRect.y + Number.parseInt(to[1].replace('px', '')) + dy));
+          this._circle.setAttribute('cx', <any>(rect.x - this.designerView.containerBoundingRect.x + Number.parseFloat(to[0].replace('px', '')) + dx));
+          this._circle.setAttribute('cy', <any>(rect.y - this.designerView.containerBoundingRect.y + Number.parseFloat(to[1].replace('px', '')) + dy));
+          this._circle2.setAttribute('cx', <any>(rect.x - this.designerView.containerBoundingRect.x + Number.parseFloat(to[0].replace('px', '')) + dx));
+          this._circle2.setAttribute('cy', <any>(rect.y - this.designerView.containerBoundingRect.y + Number.parseFloat(to[1].replace('px', '')) + dy));
         }
         break;
       case EventNames.PointerUp:
@@ -54,8 +54,8 @@ export class TransformOriginExtension extends AbstractExtension {
         if (this._startPos) {
           const dx = event.x - this._startPos.x;
           const dy = event.y - this._startPos.y;
-          const x = Number.parseInt(to[0].replace('px', ''));
-          const y = Number.parseInt(to[1].replace('px', ''));
+          const x = Number.parseFloat(to[0].replace('px', ''));
+          const y = Number.parseFloat(to[1].replace('px', ''));
           const newX = (dx + x);
           const newY = (dy + y);
           const przX = Math.round(newX / rect.width * 10000) / 100; //round to 2 decimal places
